Moving onManageSelectTrait to traits.mjs
